:truck: Move layout realted file into a file
