+++ packages/web/src/components/Layout/Footer.tsx
